+++ src/config.ts
@@ -5,14 +5,12 @@ type Config = {
   skipVerification: boolean
   appUrl: string
   pohAppUrl: string
-  collablandEcdsaPublicKey: string
   collablandEd25519PublicKeyHex: string
 }
 
 const validationSchema = yup.object({
   loglevel: yup.string().optional().default('debug'),
   skipVerification: yup.boolean().required(),
-  collablandEcdsaPublicKey: yup.string().required(),
   collablandEd25519PublicKeyHex: yup.string().required(),
   pohAppUrl: yup.string().required(),
   appUrl: yup.string().required(),
@@ -24,7 +22,6 @@ const loadCfg = (): Config => {
   const config = validationSchema.cast({
     loglevel: process.env.LOG_LEVEL,
     skipVerification: skipVerification ? skipVerification === 'true' : false,
-    collablandEcdsaPublicKey: process.env.COLLABLAND_ECDSA_PUBLIC_KEY,
     collablandEd25519PublicKeyHex: process.env.COLLABLAND_ED25519_PUBLIC_KEY_HEX,
     pohAppUrl: process.env.POH_APP_URL,
     appUrl: process.env.APP_URL,